migrated the methods from runner.py to application.py
